Extract collapsible functions into a separate file
